# fix: remove dynamic creation of static value