## Title

Initialize static values on module

## Relevant issues

None

## Pre-Submission checklist

**Please complete all items before asking a LiteLLM maintainer to review your PR**

- [ ] I have Added testing in the [`tests/litellm/`](https://github.com/BerriAI/litellm/tree/main/tests/litellm) directory, **Adding at least 1 test is a hard requirement** - [see details](https://docs.litellm.ai/docs/extras/contributing_code)
- [ ] I have added a screenshot of my new test passing locally 
- [x] My PR passes all unit tests on [`make test-unit`](https://docs.litellm.ai/docs/extras/contributing_code)
- [x] My PR's scope is as isolated as possible, it only solves 1 specific problem


## Type

🧹 Refactoring

## Changes

Removed expensive memory operations from the hot path and placed it on the module.

## Changes
- Avoid repeated memory allocations on the hot path by initializing static arrays once at the module level.

## Performance Comparison

The following benchmarks show reduced latency and improved throughput after the change:


| Metric               | Baseline      | After Changes |
|----------------------|---------------|---------------|
| Total requests       | 33k           | 33k           |
| Failures             | 0             | 0             |
| Throughput (RPS)     | 125           | 136           |
| Median latency (p50) | 47 ms         | 38 ms         |
| P95 latency          | 240 ms        | 180 ms        |
| P99 latency          | 420 ms        | 375 ms        |
| Max latency          | 1.3 s         | 0.9 s         |
| Average latency      | 72 ms         | 55 ms         |